Fix --use correct annotations for variable

+++ 0x00-python_variable_annotations/6-sum_mixed_list.py
@@ -1,6 +1,6 @@
 #!/usr/bin/env python3
 
-def sum_mixed_list(mxd_list: list[float | int])->float:
+def sum_mixed_list(mxd_list: list[int | float])-> float:
     """Takes list of floats and ints, calculate the sum
     Args"
         mxd_list: list of floats and ints
